# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Përgjithmonë Photobook Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangePasswordResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Current password incorrect / not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── CATEGORIES ──────────────────────────────────────────────────────────
...
  /categories/{categoryId}/subcategories:
    get:
      operationId: listSubcategories
      tags: [categories]
      summary: List subcategories for a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of subcategories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subcategory"

  # ─── TEMPLATES ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, email, role, emailVerified, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        phone:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        emailVerified:
          type: boolean
        lastLoginAt:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [user, accessToken]
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
        recaptchaToken:
          type: string
    LoginInput:
      type: object
      required: [password]
      properties:
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
        recaptchaToken:
          type: string
...
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
    VerifyEmailInput:
      type: object
      required: [token]
      properties:
        token:
          type: string

    # ── Categories ────────────────────────────────────────────────────────
...
    Subcategory:
      type: object
      required: [id, categoryId, slug, nameAl, nameEn, sortOrder, isActive]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        slug:
          type: string
        nameAl:
          type: string
        nameEn:
          type: string
        previewImage:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
...
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    ChangePasswordResult:
      type: object
      required: [success, accessToken]
      properties:
        success:
          type: boolean
        accessToken:
          type: string
...